+++ .circleci/config.yml
@@ -155,27 +155,3 @@ workflows:
           name: integrations-py312
           python-version: '3.12'
           task: integrations
-      - test:
-          requires:
-            - check_for_changes
-          name: asyncio-py39
-          python-version: '3.9'
-          task: async
-      - test:
-          requires:
-            - check_for_changes
-          name: asyncio-py310
-          python-version: '3.10'
-          task: async
-      - test:
-          requires:
-            - check_for_changes
-          name: asyncio-py311
-          python-version: '3.11'
-          task: async
-      - test:
-          requires:
-            - check_for_changes
-          name: asyncio-py312
-          python-version: '3.12'
-          task: async
