+++ docs/reference/drivers/AsyncDriver.rst
@@ -2,7 +2,7 @@ AsyncDriver
 ___________
 Use this driver in an async context. E.g. for use with FastAPI.
 
-.. autoclass:: hamilton.experimental.h_async.AsyncDriver
+.. autoclass:: hamilton.async_driver.AsyncDriver
    :special-members: __init__
    :members:
 
@@ -11,6 +11,6 @@ Async Builder
 
 Builds a driver in an async context -- use ``await builder....build()``.
 
-.. autoclass:: hamilton.experimental.h_async.Builder
+.. autoclass:: hamilton.async_driver.Builder
    :special-members: __init__
    :members:
