+++ docs/reference/graph-adapters/AsyncGraphAdapter.rst
@@ -3,7 +3,7 @@ h_async.AsyncGraphAdapter
 =========================
 
 
-.. autoclass:: hamilton.experimental.h_async.AsyncGraphAdapter
+.. autoclass:: hamilton.async_driver.AsyncGraphAdapter
    :special-members: __init__
    :members:
    :inherited-members:
