--- plugin_tests/h_async/conftest.py
@@ -1,4 +1,0 @@
-from hamilton import telemetry
-
-# disable telemetry for all tests!
-telemetry.disable_telemetry()
