--- plugin_tests/h_async/requirements-test.txt
@@ -1,1 +1,0 @@
-pytest-asyncio
