+++ tests/resources/simple_async_module.py
