+++ tests/test_async_driver.py
@@ -4,8 +4,7 @@ from unittest import mock
 import pandas as pd
 import pytest
 
-from hamilton import base
-from hamilton.experimental import h_async
+from hamilton import async_driver, base
 from hamilton.lifecycle.base import (
     BasePostGraphConstruct,
     BasePostGraphConstructAsync,
@@ -30,36 +29,36 @@ async def async_identity(n: int) -> int:
 @pytest.mark.asyncio
 async def test_await_dict_of_coroutines():
     tasks = {n: async_identity(n) for n in range(0, 10)}
-    results = await h_async.await_dict_of_tasks(tasks)
+    results = await async_driver.await_dict_of_tasks(tasks)
     assert results == {n: await async_identity(n) for n in range(0, 10)}
 
 
 @pytest.mark.asyncio
 async def test_await_dict_of_tasks():
     tasks = {n: asyncio.create_task(async_identity(n)) for n in range(0, 10)}
-    results = await h_async.await_dict_of_tasks(tasks)
+    results = await async_driver.await_dict_of_tasks(tasks)
     assert results == {n: await async_identity(n) for n in range(0, 10)}
 
 
 # The following are not parameterized as we need to use the event loop -- fixtures will complicate this
 @pytest.mark.asyncio
 async def test_process_value_raw():
-    assert await h_async.process_value(1) == 1
+    assert await async_driver.process_value(1) == 1
 
 
 @pytest.mark.asyncio
 async def test_process_value_coroutine():
-    assert await h_async.process_value(async_identity(1)) == 1
+    assert await async_driver.process_value(async_identity(1)) == 1
 
 
 @pytest.mark.asyncio
 async def test_process_value_task():
-    assert await h_async.process_value(asyncio.create_task(async_identity(1))) == 1
+    assert await async_driver.process_value(asyncio.create_task(async_identity(1))) == 1
 
 
 @pytest.mark.asyncio
 async def test_driver_end_to_end():
-    dr = h_async.AsyncDriver({}, simple_async_module)
+    dr = async_driver.AsyncDriver({}, simple_async_module)
     all_vars = [var.name for var in dr.list_available_variables() if var.name != "return_df"]
     result = await dr.raw_execute(final_vars=all_vars, inputs={"external_input": 1})
     result["a"] = result["a"].to_dict()  # convert to dict for comparison
@@ -85,7 +84,7 @@ async def test_driver_end_to_end():
 @mock.patch("hamilton.telemetry.send_event_json")
 @mock.patch("hamilton.telemetry.g_telemetry_enabled", True)
 async def test_driver_end_to_end_telemetry(send_event_json):
-    dr = h_async.AsyncDriver({}, simple_async_module, result_builder=base.DictResult())
+    dr = async_driver.AsyncDriver({}, simple_async_module, result_builder=base.DictResult())
     with mock.patch("hamilton.telemetry.g_telemetry_enabled", False):
         # don't count this telemetry tracking invocation
         all_vars = [var.name for var in dr.list_available_variables() if var.name != "return_df"]
@@ -156,7 +155,7 @@ async def test_async_driver_end_to_end_async_lifecycle_methods():
 
     adapter = AsyncTrackingAdapter(tracked_calls)
 
-    dr = await h_async.AsyncDriver(
+    dr = await async_driver.AsyncDriver(
         {}, simple_async_module, result_builder=base.DictResult(), adapters=[adapter]
     ).ainit()
     all_vars = [var.name for var in dr.list_available_variables() if var.name != "return_df"]
@@ -220,7 +219,7 @@ async def test_async_driver_end_to_end_sync_lifecycle_methods():
 
     adapter = AsyncTrackingAdapter(tracked_calls)
 
-    dr = await h_async.AsyncDriver(
+    dr = await async_driver.AsyncDriver(
         {}, simple_async_module, result_builder=base.DictResult(), adapters=[adapter]
     ).ainit()
     all_vars = [var.name for var in dr.list_available_variables() if var.name != "return_df"]
